requirements.txt
# Core Dependencies
streamlit>=1.28.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.20
streamlit-agraph>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
pylint>=3.0.0
mypy>=1.7.0

